FT2023-60:make individual file for  each class and start changes according to the review.
