+++ html/classes/ValidateOtp.php
@@ -1,0 +1,16 @@
+<?php
+class ValidateOtp
+{
+
+  public $Err = "";
+  public function __construct($otp)
+  {
+    if ($otp == $_SESSION["otp"]) {
+      unset($_SESSION["otp"]);
+      header('location:reset.php');
+    } else {
+      $this->Err = "* Enter Correct OTP";
+    }
+  }
+}
+?>
